requirements.txt
flask
flask-cors
openai
pinecone-client
sentence-transformers
python-dotenv
langchain
langchain-pinecone
